Modification des pontons et du spawn des players
Les pontons ont le tags floor, donc on peut poser des pièges dessus (j'ai bien test ca marche). Normalement les joueurs spawn plus haut et sont setactive false plus tôt, donc ils ne devraient plus prendre la balle en spawnant (j'ai test et pas eu le problème, mais c'est pas dit que c'est 100% réglé)

+++ Assets/Scripts/Health/HealthScript.cs
@@ -47,7 +47,7 @@ namespace PUNTutorial
 
         //fonction pour heal la barre de vie au dessus du joueur
         [PunRPC]
-        public void Heal(float heal)
+        public void Heal(int heal)
         {
             HealthSlider.value = HealthSlider.value + heal;
         }
